PYTHON-6017 Consolidate QE enums into shared class
